demo-app/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
distro==1.9.0
docstring_parser==0.16
fastapi==0.115.5
Flask==2.1.0
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-discoveryengine==0.13.4
google-cloud-resource-manager==1.13.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.13.1
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.0
Markdown==3.7
MarkupSafe==3.0.2
numpy==2.1.3
openai==1.55.3
openparse==0.7.0
packaging==24.2
pdfminer.six==20240706
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.29.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
PyMuPDF==1.24.14
pypdf==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
tiktoken==0.8.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.32.1
vertexai==1.71.1
watchfiles==1.0.0
websockets==14.1
Werkzeug==2.1.0
